Week1-SQL/requirements.txt
faker>=13.0.0
psycopg2>=2.9
python-dotenv>=1.0